NAV-25849: Lager en egen "avslagperiode" pakke
