--- src/main/kotlin/no/nav/familie/ks/sak/kjerne/behandling/steg/vedtak/vedtaksperiode/Avslagsperiode.kt
@@ -1,9 +1,0 @@
-package no.nav.familie.ks.sak.kjerne.behandling.steg.vedtak.vedtaksperiode
-
-import java.time.LocalDate
-
-data class Avslagsperiode(
-    override val periodeFom: LocalDate?,
-    override val periodeTom: LocalDate?,
-    override val vedtaksperiodetype: Vedtaksperiodetype = Vedtaksperiodetype.AVSLAG,
-) : Vedtaksperiode
